[BEIN-14712]Update environment configurations and Android build.gradle file
This commit updates the environment configurations of the project by changing the BUNDLE_IDENTIFIER and BUNDLE_DISPLAY_NAME values. It also updates the values of several other environment variables. Additionally, the Android build.gradle file was updated to read the values of these environment variables and use them in the configuration of the project. The new configuration values were also set for the 'development' product flavor in the file.